WebSockets/tests/submissions/Opera/interfaces/CloseEvent/003.html

<!doctype html>
<title>WebSockets: dispatching a CloseEvent</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../constants.js></script>
<div id=log></div>
<script>
setTestTimeout(2000);
async_test(function(t) {
  var e = document.createEvent('CloseEvent');
  e.initCloseEvent('foo', true, true, true);
  var div = document.createElement('div');
  div.addEventListener('foo', function(ev) {
    assert_equals(ev.bubbles, true);
    assert_equals(ev.cancelable, true);
    assert_equals(ev.wasClean, true);
    end();
  }, false);
  div.dispatchEvent(e);
});
</script>
